externalize more configs, cleaned up event utils, created utils module
